Move stripeui to vuetifyx
